requirements.txt
Django==3.2.10
django-environ==0.4
psycopg2-binary==2.9.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.13.0
django-allauth==0.50.0
gunicorn==20.1.0
django-celery-results==2.4.0
python-gettext==4.0
django-filter==22.1
celery[redis]==5.2.7
redis==4.3.4
Pillow==9.2.0
requests==2.28.1